Separated input/output threads from analysis threads by forking the process.

+++ src/config.h
@@ -42,9 +42,9 @@
 
 typedef struct _MLE_STATS_
 {
-    int input;
-    int analysis;
-    int output;
+    unsigned long input;
+    unsigned long analysis;
+    unsigned long output;
 } MLE_STATS;
 
 typedef struct _INPUT_CONFIG_
